+++ src/services/userService.js
@@ -33,9 +33,11 @@ export const deleteUser = async (userId) => {
             throw new Error('Usuario no encontrado')
         }
 
-        await Cart.findByIdAndDelete(user.cart)
-        await userRepository.deleteUser(userId)
-        return new UserDTO(user)
+        if (user.role !== 'admin') {
+            await Cart.findByIdAndDelete(user.cart)
+            await userRepository.deleteUser(userId)
+            return new UserDTO(user)
+        }
     } catch (error) {
         throw new Error('Error al eliminar el perfil: ' + error.message)
     }
